:hammer: Move Searchable up a level
