+++ src/rogallo/widgets/viewer/gemtext/__init__.py
@@ -26,7 +26,6 @@ from .list_item import GemtextListItem
 from .paragraph import GemtextParagraph
 from .preformatted import GemtextPreformatted
 from .quote import GemtextQuote
-from .searchable import Searchable
 from .text import GemtextText
 
 ##############################################################################
@@ -70,7 +69,6 @@ __all__ = [
     "GemtextContent",
     "GemtextLink",
     "get_block_widget",
-    "Searchable",
 ]
 
 ### __init__.py ends here
