+++ src/rogallo/widgets/viewer/searchable.py
